+++ kalaha/config.py
@@ -1,2 +1,2 @@
 __app_name__ = "Kalaha"
-__version__ = "0.2.0"
+__version__ = "0.3.0"