+++ kalaha/controlers/__init__.py
@@ -1,0 +1,1 @@
+from .ClientController import ClientController