+++ kalaha/widgets/__init__.py
@@ -2,3 +2,5 @@ from .PitButtons import PitButton, BaseButton
 from .BoardWidget import BoardWidget
 from .InfoWidget import InfoWidget
 from .StatusWidget import StatusWidget
+from .JoinBoardWidget import JoinBoardWidget
+from .CreateBoardWidget import CreateBoardWidget