+++ kalaha/windows/BoardWindow.py
@@ -1,24 +1,23 @@
 from qtpy.QtWidgets import (QDialogButtonBox, QMenu, QWidgetAction, QAction)
-from qtpy.QtCore import Slot, QTimer, QThread
+from qtpy.QtCore import Slot, QTimer, QThread, Signal
 from qrainbowstyle.windows import (FramelessWindow, FramelessWarningMessageBox, FramelessInformationMessageBox,
                                    FramelessCriticalMessageBox)
 from qrainbowstyle.widgets import StylePickerHorizontal
 
 import logging
 
-from kalaha.network import KalahaClient
 from kalaha.models import Board, AutoPlayer
-from kalaha.widgets import BoardWidget, InfoWidget, StatusWidget
+from kalaha.widgets import BoardWidget, InfoWidget, StatusWidget, JoinBoardWidget, CreateBoardWidget
 
 
 class BoardWindow(FramelessWindow):
+    closed = Signal()
 
     def __init__(self, parent=None):
         super(BoardWindow, self).__init__(parent)
         self.logger = logging.getLogger(self.__class__.__name__)
 
         self.player_number = 0
-        self.client = KalahaClient()
         self.auto_player_worker: AutoPlayer = None
         self.auto_player_thread: QThread = None
         self.timeout = 0
@@ -26,27 +25,6 @@ class BoardWindow(FramelessWindow):
 
         self.timer = QTimer(self)
         self.timer.timeout.connect(self.on_timer_timeout)
-
-        # connection
-        self.client.connected.connect(self.on_client_connected)
-        self.client.failed_to_connect.connect(self.on_failed_to_connect)
-
-        # board changes
-        self.client.invalid_move.connect(self.on_invalid_move)
-        self.client.update_board.connect(self.on_update_board)
-        self.client.setup_board.connect(self.on_setup_board)
-        self.client.your_move.connect(self.on_your_move)
-        self.client.turn_timeout.connect(self.on_turn_timeout)
-
-        # opponent status
-        self.client.opponent_connected.connect(self.on_opponent_connected)
-        self.client.opponent_not_connected.connect(lambda: self.show_info_dialog("Opponent not connected!"))
-        self.client.opponent_disconnected.connect(self.on_opponent_disconnected)
-
-        # game result
-        self.client.you_won.connect(lambda: self.on_game_result("You won!"))
-        self.client.you_lost.connect(lambda: self.on_game_result("You lost!"))
-        self.client.you_tied.connect(lambda: self.on_game_result("Draw!"))
 
         # windows
         self.message_box = None
@@ -57,10 +35,22 @@ class BoardWindow(FramelessWindow):
         self.addContentWidget(self.status_widget)
 
         self.board_widget = BoardWidget(self)
+        self.board_widget.hide()
         self.addContentWidget(self.board_widget)
 
         self.info_widget = InfoWidget(self)
+        self.info_widget.hide()
         self.addContentWidget(self.info_widget)
+
+        self.join_widget = JoinBoardWidget(self)
+        self.join_widget.create_button.clicked.connect(self.set_board_create_visible)
+        self.join_widget.hide()
+        self.addContentWidget(self.join_widget)
+
+        self.create_widget = CreateBoardWidget(self)
+        self.create_widget.back_button.clicked.connect(self.set_board_create_visible)
+        self.create_widget.hide()
+        self.addContentWidget(self.create_widget)
 
         # other widgets
         self.menu = QMenu(self)
@@ -78,22 +68,20 @@ class BoardWindow(FramelessWindow):
         self.addMenu(self.menu)
 
     @Slot(str, int)
-    def connect_to_server(self, host: str, port: int):
-        self.logger.debug(f"Connecting to server: {host}:{port}")
-        self.client.start(host, port)
-
-    @Slot(str, int)
     def on_client_connected(self, host: str, port: int):
-        self.logger.info(f"Connected to {host}:{port}")
+        self.status_widget.hide()
+        self.join_widget.show()
         self.status_widget.set_status("Waiting for opponent...")
-        self.board_widget.pit_clicked.connect(self.on_pit_clicked)
 
     @Slot(str, int)
     def on_failed_to_connect(self, host: str, port: int):
+        self.show_info_dialog("Failed to connect!")
         self.status_widget.set_status(f"Failed to connect to {host}:{port}")
 
     @Slot()
     def on_opponent_connected(self):
+        self.create_widget.hide()
+        self.join_widget.hide()
         self.status_widget.hide()
         self.board_widget.show()
         self.info_widget.show()
@@ -101,8 +89,6 @@ class BoardWindow(FramelessWindow):
     @Slot()
     def on_opponent_disconnected(self):
         self.timer.stop()
-        self.auto_player_worker.stop.emit()
-        self.client.disconnect_from_server()
         self.logger.debug("Opponent disconnected!")
         if self.message_box:
             self.message_box.close()
@@ -113,11 +99,26 @@ class BoardWindow(FramelessWindow):
         self.message_box.setText("Opponent disconnected! Click OK to close.")
         self.message_box.show()
 
+    @Slot()
+    def set_board_create_visible(self):
+        if self.join_widget.isVisible():
+            self.join_widget.hide()
+            self.create_widget.show()
+        else:
+            self.create_widget.hide()
+            self.join_widget.show()
+
+    @Slot(int)
+    def on_board_joined(self, id: int):
+        self.join_widget.hide()
+        self.create_widget.hide()
+        self.status_widget.show()
+
     # Board events
     @Slot(Board, list, int)
     def on_setup_board(self, board, allowed_pits, player_number):
         """
-        Setup the board and pits
+        Set up the board and pits
         """
         self.board = board.board
         self.player_number = player_number
@@ -143,15 +144,6 @@ class BoardWindow(FramelessWindow):
         if your_move and self.auto_player_checkbox.isChecked():
             self.auto_player_worker.calculate_move.emit(self.board, self.player_number)
         self.show_info_dialog(message)
-
-    @Slot(int)
-    def on_pit_clicked(self, pit_index: int):
-        if not self.client.is_running():
-            return
-        if self.auto_player_checkbox.isChecked():
-            self.show_info_dialog("Auto play is on!")
-            return
-        self.client.make_move(pit_index)
 
     @Slot(str)
     def on_invalid_move(self, message: str):
@@ -215,29 +207,16 @@ class BoardWindow(FramelessWindow):
         self.message_box.show()
 
     @Slot()
-    def show(self):
-        """
-        Show window with spinner.
-        """
-        self.info_widget.hide()
-        self.board_widget.hide()
-        self.status_widget.show()
-        super(BoardWindow, self).show()
-
-    @Slot()
     def close(self):
         """
         Close the window and message box if it is open.
         Disconnect from the server and wait for the thread to finish.
         If auto player is running, stop it and wait for the thread to finish.
         """
-        self.hide()
         if self.message_box is not None:
             self.message_box.close()
-        if self.client.is_running():
-            self.client.close()
-            self.client.wait()
         if self.auto_player_thread:
             self.auto_player_thread.quit()
             self.auto_player_thread.wait()
-        super(BoardWindow, self).close()
+        super().close()
+        self.closed.emit()