+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+- 0.3.0:
+  - Create ClientController
+  - Add CMaximax algorithm
+  - Add option to create multiple boards secured by PIN
+  - Use Python 3.10 feature - pattern matching
+  - Change Dockerfile base image to `python:3.10-slim`
+  - Optimize game engine code
+  - Use `QBalancedServer` instead of `QThreadedServer`
+
 - 0.2.0:
   - Rename CAutoPlayer to CMinimax
   - Fix checking if game over in AutoPlayer